# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: QWIZDS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ai/generate-quiz:
    post:
      operationId: generateQuiz
      tags: [ai]
      summary: AI-generate a full quiz from topic or text
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateQuizBody"
      responses:
        "200":
          description: Generated quiz
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratedQuiz"
  /ai/improve-question:
    post:
      operationId: improveQuestion
      tags: [ai]
      summary: AI-improve an existing question
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImproveQuestionBody"
      responses:
        "200":
          description: Improved question
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImprovedQuestion"
  /ai/chat:
    post:
      operationId: aiChat
      tags: [ai]
      summary: Chat with AI assistant about quizzes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiChatBody"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiChatResponse"
...
components:
  schemas:
...
    GenerateQuizBody:
      type: object
      properties:
        topic:
          type: string
        text:
          type: string
        questionCount:
          type: integer
        difficulty:
          type: string
          enum: [easy, medium, hard]
        questionTypes:
          type: array
          items:
            type: string
      required:
        - topic
    GeneratedQuiz:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        questions:
          type: array
          items:
            $ref: "#/components/schemas/GeneratedQuestion"
      required:
        - title
        - description
        - questions
    GeneratedQuestion:
      type: object
      properties:
        questionText:
          type: string
        questionType:
          type: string
          enum: [multiple_choice, true_false, short_answer]
        options:
          type: array
          items:
            type: string
        correctAnswer:
          type: string
        explanation:
          type: string
        timeLimit:
          type: integer
        points:
          type: integer
      required:
        - questionText
        - questionType
        - options
        - correctAnswer
        - explanation
        - timeLimit
        - points
    ImproveQuestionBody:
      type: object
      properties:
        questionText:
          type: string
        options:
          type: array
          items:
            type: string
        correctAnswer:
          type: string
        feedback:
          type: string
      required:
        - questionText
        - correctAnswer
    ImprovedQuestion:
      type: object
      properties:
        questionText:
          type: string
        options:
          type: array
          items:
            type: string
        correctAnswer:
          type: string
        explanation:
          type: string
        improvements:
          type: string
      required:
        - questionText
        - options
        - correctAnswer
        - explanation
        - improvements
    AiChatBody:
      type: object
      properties:
        message:
          type: string
        context:
          type: string
      required:
        - message
    AiChatResponse:
      type: object
      properties:
        reply:
          type: string
      required:
        - reply
...